+++ src/Symfony/Bundle/SecurityBundle/DataCollector/SecurityDataCollector.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Bundle\SecurityBundle\DataCollector;
 
+use Symfony\Bundle\SecurityBundle\Debug\TraceableFirewallListener;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Role\Role;
 use Symfony\Component\Security\Core\Role\RoleHierarchyInterface;
@@ -38,6 +39,7 @@ class SecurityDataCollector extends DataCollector implements LateDataCollectorIn
     private $logoutUrlGenerator;
     private $accessDecisionManager;
     private $firewallMap;
+    private $firewall;
     private $hasVarDumper;
 
     /**
@@ -48,14 +50,16 @@ class SecurityDataCollector extends DataCollector implements LateDataCollectorIn
      * @param LogoutUrlGenerator|null             $logoutUrlGenerator
      * @param AccessDecisionManagerInterface|null $accessDecisionManager
      * @param FirewallMapInterface|null           $firewallMap
+     * @param TraceableFirewallListener|null      $firewall
      */
-    public function __construct(TokenStorageInterface $tokenStorage = null, RoleHierarchyInterface $roleHierarchy = null, LogoutUrlGenerator $logoutUrlGenerator = null, AccessDecisionManagerInterface $accessDecisionManager = null, FirewallMapInterface $firewallMap = null)
+    public function __construct(TokenStorageInterface $tokenStorage = null, RoleHierarchyInterface $roleHierarchy = null, LogoutUrlGenerator $logoutUrlGenerator = null, AccessDecisionManagerInterface $accessDecisionManager = null, FirewallMapInterface $firewallMap = null, TraceableFirewallListener $firewall = null)
     {
         $this->tokenStorage = $tokenStorage;
         $this->roleHierarchy = $roleHierarchy;
         $this->logoutUrlGenerator = $logoutUrlGenerator;
         $this->accessDecisionManager = $accessDecisionManager;
         $this->firewallMap = $firewallMap;
+        $this->firewall = $firewall;
         $this->hasVarDumper = class_exists(ClassStub::class);
     }
 
@@ -157,6 +161,12 @@ class SecurityDataCollector extends DataCollector implements LateDataCollectorIn
                     'listeners' => $firewallConfig->getListeners(),
                 );
             }
+        }
+
+        // collect firewall listeners information
+        $this->data['listeners'] = array();
+        if ($this->firewall) {
+            $this->data['listeners'] = $this->firewall->getWrappedListeners();
         }
     }
 
@@ -294,6 +304,11 @@ class SecurityDataCollector extends DataCollector implements LateDataCollectorIn
     public function getFirewall()
     {
         return $this->data['firewall'];
+    }
+
+    public function getListeners()
+    {
+        return $this->data['listeners'];
     }
 
     /**